agent-systems/igwe/requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
jinja2>=3.1.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
psycopg2-binary>=2.9.9

# Task queue
celery>=5.3.6
redis>=5.0.1

# Communication channels
sendgrid>=6.11.0
twilio>=8.11.0

# HTTP requests (async support for Apify)
httpx>=0.26.0
requests>=2.31.0

# LLM integration
openai>=1.10.0

# Calendar
pytz>=2024.1
python-dateutil>=2.8.2

# From Lead_Pipeline (reuse)
pandas>=1.3.5
numpy>=1.20.0
beautifulsoup4>=4.10.0
html2text>=2020.1.16
fake-useragent>=0.1.11
selenium>=4.1.0
webdriver-manager>=3.5.4
nltk>=3.6.7
tenacity>=8.0.1

# Utils
loguru>=0.7.2
